Terrible refactoring

+++ BookAccounting/Data/Repositories/IssuedBooksRepository.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using BookAccounting.Data.Models;
+using Microsoft.Data.Sqlite;
+using Microsoft.EntityFrameworkCore;
 
 namespace BookAccounting.Data.Repositories
 {
@@ -14,7 +16,7 @@ namespace BookAccounting.Data.Repositories
         {
             using (var db = new LibraryDbContext())
             {
-                return db.IssuedBooks.ToList();
+                return db.IssuedBooks.Include(ib => ib.Book).Include(ib => ib.Reader).ToList();
             }
         }
 
@@ -22,8 +24,17 @@ namespace BookAccounting.Data.Repositories
         {
             using (var db = new LibraryDbContext())
             {
+                var book = db.Books.FirstOrDefault(b => b.Id == entity.IdBook);
+
+                if (book == null)
+                    throw new SqliteException("book doesn't exists in database", 1);
+                else if (book.Count <= 0)
+                    throw new SqliteException("books count <= 0", 1);
+
+                book.Count--;
                 var entry = db.IssuedBooks.Add(entity);
                 db.SaveChanges();
+
                 return entry.Entity;
             }
         }
@@ -49,6 +60,12 @@ namespace BookAccounting.Data.Repositories
         {
             using (var db = new LibraryDbContext())
             {
+                var book = db.Books.FirstOrDefault(b => b.Id == entity.IdBook);
+
+                if (book == null)
+                    throw new SqliteException("book doesn't exists in database", 1);
+
+                book.Count++;
                 db.IssuedBooks.Remove(entity);
                 db.SaveChanges();
             }
